[Erlangga] ignore previous commit

+++ db/migrate/20120813111226_add_rating_to_activities.rb
@@ -1,5 +1,5 @@
 class AddRatingToActivities < ActiveRecord::Migration
   def change
-    add_column :activities, :rating, :double
+    add_column :activities, :rating, :float
   end
 end
